Add Electron setup for desktop app
