+++ index.html
@@ -38,6 +38,6 @@
         </div>
     </div>
     <script type="module" src="soundtouch.js"></script>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="renderer.js"></script>
 </body>
 </html>
